big (>2GB) AVI files support - patch by Wolfram Gloger <wg@malloc.de>
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@6057 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/aviheader.h
@@ -89,10 +89,10 @@ typedef struct {
   // index stuff:
   void* idx;
   int idx_size;
-  int idx_pos;
-  int idx_pos_a;
-  int idx_pos_v;
-  int idx_offset;  // ennyit kell hozzaadni az index offset ertekekhez
+  off_t idx_pos;
+  off_t idx_pos_a;
+  off_t idx_pos_v;
+  off_t idx_offset;  // ennyit kell hozzaadni az index offset ertekekhez
   // interleaved PTS stuff:
   int skip_video_frames;
   int audio_streams;
